"Refactored delete modals in Customer, CustomerOrders, and Products views to use a partial view and a view model; updated corresponding code in BasePageModel"
This commit refactors the delete modals in the Customer, CustomerOrders, and Products views to use a partial view and a view model, and updates the corresponding code in BasePageModel. The delete modals are now defined in a separate partial view called "_DeleteDialogModal" and are passed a view model called "DeleteModalViewModel" containing the necessary properties for the delete modal. This refactoring improves code reusability and separation of concerns.